linux-server/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
retrying>1.3.0,<1.4.0
cloudshell-cli>=4.0.0,<5.0.0
